modules: environmental: channel def in module header
Environmental channel definition into dedicated header.
