+++ tests/module/cloud/CMakeLists.txt
@@ -24,6 +24,7 @@ zephyr_include_directories(../../../app/src/modules/cloud)
 zephyr_include_directories(../../../app/src/common)
 zephyr_include_directories(../../../app/src/modules/battery)
 zephyr_include_directories(../../../app/src/modules/network)
+zephyr_include_directories(../../../app/src/modules/environmental)
 zephyr_include_directories(${NRF_DIR}/subsys/net/lib/nrf_cloud/include)
 zephyr_include_directories(${NRF_DIR}/../modules/lib/cjson)
