+++ tests/module/environmental/CMakeLists.txt
@@ -25,6 +25,7 @@ zephyr_include_directories(${ZEPHYR_BASE}/include/zephyr/)
 zephyr_include_directories(${ZEPHYR_BASE}/subsys/testsuite/include)
 zephyr_include_directories(../../../app/src/common)
 zephyr_include_directories(../../../app/src/modules/cloud)
+zephyr_include_directories(../../../app/src/modules/environmental)
 
 target_link_options(app PRIVATE --whole-archive)
 # Options that cannot be passed through Kconfig fragments
